CRM-80 MediaBundle - Secteur  logement photos  finis

+++ src/Mondofute/Bundle/FournisseurBundle/Controller/FournisseurController.php
@@ -445,8 +445,8 @@ class FournisseurController extends Controller
                 }
             }
             if (!$interlocuteurController->testInterlocuteursLoginExist($fournisseur->getInterlocuteurs())) {
-
-                $this->mAJSites($fournisseur);
+//
+//                $this->mAJSites($fournisseur);
 
                 foreach ($fournisseur->getInterlocuteurs() as $interlocuteur) {
                     $interlocuteur->setFournisseur($fournisseur);
@@ -455,7 +455,10 @@ class FournisseurController extends Controller
                 $em->persist($fournisseur);
                 $em->flush();
 
-                if ($originalLogo != $fournisseur->getLogo()) {
+
+                $this->mAJSites($fournisseur);
+
+                if (!empty($originalLogo) && $originalLogo != $fournisseur->getLogo() ) {
                     $em->remove($originalLogo);
                     $em->flush();
                 }
